feat: move models to new domain/ folder
