+++ src/app/adapters/secondary/spotify/repositories/spotify-playlists.repository.ts
@@ -1,7 +1,7 @@
 import Spotify from 'spotify-web-api-js';
 import { Injectable, inject } from '@angular/core';
-import { Playlist } from 'src/app/models/playlist.model';
 import { SpotifyHelpers } from '../helpers/spotify.helpers';
+import { Playlist } from 'src/app/domain/models/playlist.model';
 import { SPOTIFY_API_PROVIDER } from '../providers/spotify-api.provider';
 import { SessionService } from 'src/app/application/services/session.service';
 import { PlaylistsRepository } from 'src/app/repositories/playlists.repository';
