+++ src/app/common/factories.ts
@@ -1,6 +1,6 @@
-import { Music } from 'src/app/models/music.model';
-import { Artist } from 'src/app/models/artist.model';
-import { Playlist } from 'src/app/models/playlist.model';
+import { Music } from 'src/app/domain/models/music.model';
+import { Artist } from 'src/app/domain/models/artist.model';
+import { Playlist } from 'src/app/domain/models/playlist.model';
 
 export function newArtist(): Artist {
   return {
