+++ src/app/domain/models/artist.model.ts
