+++ src/app/repositories/musics.repository.ts
@@ -1,5 +1,5 @@
 import { InjectionToken } from '@angular/core';
-import { Music } from 'src/app/models/music.model';
+import { Music } from 'src/app/domain/models/music.model';
 
 export interface MusicsRepository {
   getMusics(offset: number, limit: number): Promise<Music[]>;
